能够根据位置正反漏沙

+++ Sources/MCU6050.h
@@ -46,6 +46,6 @@ uchar I2C_RecvByte(void);
 uchar Single_ReadI2C(uchar REG_Address);						//读取I2C数据
 void  Single_WriteI2C(uchar REG_Address,uchar REG_data);	//向I2C写入数据
 int GetData(uchar REG_Address);
-short MPU6050_GetAngle(float x, float y,float z,uint dir);
+int MPU6050_GetAngle(float x, float y,float z,uint dir);
 
 #endif
